Split config token and TLS helpers
